+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Updated
 
 * Better exception for smart enum errors
+* Sponsorship request
 
 ## [0.25.0] - 12/28/2023